Import and edit CSV
I had to import a csv file and use textblob which finds the polarity and be able to decide what the data was telling me from the imported csv file. I edited the csv file by writing a line of code that would allow me read it and have the data go into a list I created.